Updated architecture to a sensible one.
Previously, everything just worked using global state. That was just silly. So
now, the old API still works, but you can create an instance of the Profanalyzer
class with all the same methods as the old API, but all changes are local to the
object in question.

Also, there was a typo: Profanalyzer.subtitutions= . That's just embarrassing.